+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add Yeoman instructions @robgietema
 - Add files and images loading in profiles @robgietema
 - Add relation list fields @robgietema
+- Add @export view to export current content object data @robgietema
 
 ### Bugfix